move page-not-found component to pages folder | stylize not found page
