+++ src/app/app-routing.module.ts
@@ -6,7 +6,7 @@ import { OeisComponent } from './components/oeis/oeis.component';
 import { ProjectsComponent } from './components/projects/projects.component';
 import { QuotesComponent } from './components/quotes/quotes.component';
 import { MainPageComponent } from './pages/main-page/main-page.component';
-import { PageNotFoundComponent } from "./components/page-not-found/page-not-found.component";
+import { PageNotFoundComponent } from "./pages/page-not-found/page-not-found.component";
 import { PdfFolderIndexComponent } from "./pages/pdf-folder-index/pdf-folder-index.component";
 
 const routes: Routes = [
