+++ src/app/pages/page-not-found/page-not-found.component.scss
@@ -1,0 +1,14 @@
+
+:host {
+  position: absolute;
+  display: flex;
+  width: 100%;
+  min-height: 100%;
+  flex-direction: column;
+  background-color: rgb(255, 254, 194);
+}
+
+p {
+  margin: auto;
+  font-size: 50px;
+}
